pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["Джунусходжаев Амирхан <a.dzhunushodzhaev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
celery = "^5.2.6"
python-dotenv = "^0.20.0"
redis = "^4.2.2"
gunicorn = "^20.1.0"
flower = "^1.0.0"
requests = "^2.27.1"
Pillow = "^9.1.0"
GitPython = "^3.1.27"
fire = "^0.4.0"
djangorestframework = "^3.13.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.20.0"
whitenoise = "^6.1.0"
result = "^0.8.0"
self = "^2020.12.3"
django-cors-headers = "^3.12.0"
pylint = "^2.14.1"
psycopg2 = "^2.9.5"
poetry = "^1.4.1"
logger = "^1.4"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
